requirements.txt
fastapi
uvicorn
pyjwt
requests
sqlalchemy
environs